Refs #15840. Checkpointing progress on refactoring.
