+++ Framework/DataHandling/inc/MantidDataHandling/LoadEventNexus.h
@@ -5,6 +5,9 @@
 // Includes
 //----------------------------------------------------------------------
 #include "MantidAPI/IFileLoader.h"
+#include "MantidDataHandling/BankPulseTimes.h"
+#include "MantidDataHandling/ProcessBankData.h"
+
 #include "MantidDataObjects/EventWorkspace.h"
 #include <nexus/NeXusFile.hpp>
 #include <nexus/NeXusException.hpp>
@@ -18,43 +21,6 @@
 #include <mutex>
 #include <boost/lexical_cast.hpp>
 
-namespace Mantid {
-
-namespace DataHandling {
-
-/** This class defines the pulse times for a specific bank.
- * Since some instruments (ARCS, VULCAN) have multiple preprocessors,
- * this means that some banks have different lists of pulse times.
- */
-class BankPulseTimes {
-public:
-  /// Starting number for assigning periods.
-  static const unsigned int FirstPeriod;
-
-  /// Constructor with NeXus::File
-  BankPulseTimes(::NeXus::File &file, const std::vector<int> &pNumbers);
-
-  /// Constructor with vector of DateAndTime
-  BankPulseTimes(const std::vector<Kernel::DateAndTime> &times);
-
-  /// Destructor
-  ~BankPulseTimes();
-
-  /// Equals
-  bool equals(size_t otherNumPulse, std::string otherStartTime);
-
-  /// String describing the start time
-  std::string startTime;
-
-  /// Size of the array of pulse times
-  size_t numPulses;
-
-  /// Array of the pulse times
-  Kernel::DateAndTime *pulseTimes;
-
-  /// Vector of period numbers corresponding to each pulse
-  std::vector<int> periodNumbers;
-};
 
 /** @class LoadEventNexus LoadEventNexus.h Nexus/LoadEventNexus.h
