+++ Framework/DataHandling/inc/MantidDataHandling/ProcessBankData.h
@@ -1,0 +1,77 @@
+// Process bank data
+#include "MantidDataHandling/LoadEventNexus.h"
+
+//==============================================================================================
+// Class ProcessBankData
+//==============================================================================================
+// FIXME-TODO- Move ProcessBankData to new files
+/** This task does the disk IO from loading the NXS file,
+* and so will be on a disk IO mutex */
+class ProcessBankData : public Task {
+public:
+  //----------------------------------------------------------------------------------------------
+  /** Constructor
+  *
+  * @param alg :: LoadEventNexus
+  * @param entry_name :: name of the bank
+  * @param prog :: Progress reporter
+  * @param event_id :: array with event IDs
+  * @param event_time_of_flight :: array with event TOFS
+  * @param numEvents :: how many events in the arrays
+  * @param startAt :: index of the first event from event_index
+  * @param event_index :: vector of event index (length of # of pulses)
+  * @param thisBankPulseTimes :: ptr to the pulse times for this particular
+  *bank.
+  * @param have_weight :: flag for handling simulated files
+  * @param event_weight :: array with weights for events
+  * @param min_event_id ;: minimum detector ID to load
+  * @param max_event_id :: maximum detector ID to load
+  * @return
+  */
+  ProcessBankData(LoadEventNexus *alg, std::string entry_name, Progress *prog,
+                  boost::shared_array<uint32_t> event_id,
+                  boost::shared_array<float> event_time_of_flight,
+                  size_t numEvents, size_t startAt,
+                  boost::shared_ptr<std::vector<uint64_t>> event_index,
+                  boost::shared_ptr<BankPulseTimes> thisBankPulseTimes,
+                  bool have_weight, boost::shared_array<float> event_weight,
+                  detid_t min_event_id, detid_t max_event_id);
+  
+  
+  void run() override;
+
+private:
+  /// Algorithm being run
+  LoadEventNexus *alg;
+  /// NXS path to bank
+  std::string entry_name;
+  /// Vector where (index = pixel ID+pixelID_to_wi_offset), value = workspace
+  /// index)
+  const std::vector<size_t> &pixelID_to_wi_vector;
+  /// Offset in the pixelID_to_wi_vector to use.
+  detid_t pixelID_to_wi_offset;
+  /// Progress reporting
+  Progress *prog;
+  /// event pixel ID array
+  boost::shared_array<uint32_t> event_id;
+  /// event TOF array
+  boost::shared_array<float> event_time_of_flight;
+  /// # of events in arrays
+  size_t numEvents;
+  /// index of the first event from event_index
+  size_t startAt;
+  /// vector of event index (length of # of pulses)
+  boost::shared_ptr<std::vector<uint64_t>> event_index;
+  /// Pulse times for this bank
+  boost::shared_ptr<BankPulseTimes> thisBankPulseTimes;
+  /// Flag for simulated data
+  bool have_weight;
+  /// event weights array
+  boost::shared_array<float> event_weight;
+  /// Minimum pixel id
+  detid_t m_min_id;
+  /// Maximum pixel id
+  detid_t m_max_id;
+  /// timer for performance
+  Mantid::Kernel::Timer m_timer;
+}; // END-DEF-CLASS ProcessBankData
